backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Cors==3.0.10
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
gunicorn==20.1.0
requests==2.31.0
Flask-APScheduler==1.13.1
